fix: design system compliance, homepage consolidation, and search page improvements
- Fixed all design system violations on homepages (23+ hardcoded colors replaced with design tokens)
- Fixed typography: reduced H1 from lg:text-5xl to lg:text-4xl for better accessibility
- Consolidated buyer and developer homepages to 4 sections max (excluding AI helper)
- Removed 'Compare Properties' section from both homepages
- Removed filter panel borders/shadows for seamless page integration
- Fixed search page view routing: /property/search/ always shows list view (full width), /property/search-with-map/ shows map view
- Removed list view functionality entirely (grid view only)
- Fixed CSS scoping: sidebar styles only apply to map view page
- Removed shadow-md class that was causing border appearance on scroll
- Updated documentation: ui-ux-redesign-audit.md and ui-ux-redesign-status.md

+++ src/advertise/api/views/advertisement.py
@@ -36,8 +36,8 @@ class AdvertisementViewSet(viewsets.ModelViewSet):
         
         if time_spent:
             try:
-                # Convert to float and validate
-                time_spent_seconds = float(time_spent)
+                # Convert to int and validate
+                time_spent_seconds = int(time_spent)
                 if time_spent_seconds > 0:
                     ad_obj.view_time += timedelta(seconds=time_spent_seconds)
                     ad_obj.save()
